+++ discopy/pivotal.py
@@ -54,7 +54,7 @@ We also have its dagger and its transpose:
 
 from __future__ import annotations
 
-from discopy import cat, rigid, traced
+from discopy import cat, cmap, rigid, traced
 from discopy.abc import PivotalCategory
 from discopy.cat import factory
 
@@ -251,6 +251,11 @@ class Functor(rigid.Functor):
     dom = cod = Diagram
 
 
+CMap = cmap.CMap[Diagram]
+
+
+Diagram.functor_factory = Functor
+Diagram.map_factory = CMap
 Diagram.cup_factory, Diagram.cap_factory = Cup, Cap
 Id = Diagram.id